# openapi.yml
openapi: 3.0.3
info:
  title: ehsundar platform API
  version: 0.1.0
  description: Shared backend facilities (auth, profiles, settings) for every program.
paths:
...
  /api/v1/profiles/me/:
    get:
      operationId: profiles_me_retrieve
      description: Read or update the caller's profile in the current program.
      tags:
      - profiles
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
          description: ''
    patch:
      operationId: profiles_me_partial_update
      description: Read or update the caller's profile in the current program.
      tags:
      - profiles
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedPublicProfile'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedPublicProfile'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedPublicProfile'
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
          description: ''
  /api/v1/profiles/me/settings/:
    get:
      operationId: profiles_me_settings_retrieve
      description: Resolved settings for the current program, or merge overrides in.
      tags:
      - profiles
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: No response body
    patch:
      operationId: profiles_me_settings_partial_update
      description: Resolved settings for the current program, or merge overrides in.
      tags:
      - profiles
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: No response body
...
components:
  schemas:
...
    PatchedPublicProfile:
      type: object
      description: What other members of the same program may see.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        program:
          type: string
          readOnly: true
        display_name:
          type: string
          readOnly: true
        avatar_url:
          type: string
          format: uri
          readOnly: true
        bio:
          type: string
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        program:
          type: string
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        display_name:
          type: string
          maxLength: 150
        avatar_url:
          oneOf:
          - type: string
            format: uri
            maxLength: 200
          - type: string
            maxLength: 0
        bio:
          type: string
        locale:
          type: string
          maxLength: 16
        timezone:
          type: string
          maxLength: 64
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          readOnly: true
        enrolled_at:
          type: string
          format: date-time
          readOnly: true
        data: {}
        settings:
          type: object
          additionalProperties: {}
          readOnly: true
      required:
      - email
      - enrolled_at
      - id
      - program
      - role
      - settings
      - status
...
    RoleEnum:
      enum:
      - member
      - moderator
      - admin
      type: string
      description: |-
        * `member` - Member
        * `moderator` - Moderator
        * `admin` - Admin
    StatusEnum:
      enum:
      - active
      - pending
      - suspended
      type: string
      description: |-
        * `active` - Active
        * `pending` - Pending
        * `suspended` - Suspended
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT